pyproject.toml
[tool.poetry]
name = "WitnessMe"
version = "1.5.0dev"
description = "Web Inventory tool that uses Pyppeteer (headless Chrome/Chromium) and provides some extra bells & whistles to make life easier."
authors = ["Marcello Salvati <byt3bl33d3r@pm.com>"]
readme = "README.md"
homepage = "https://github.com/byt3bl33d3r/WitnessMe"
repository = "https://github.com/byt3bl33d3r/WitnessMe"
exclude = ["tests/*", "dockerfiles/*"]
include = ["LICENSE", "witnessme/signatures/*"]
license = "GPL-3.0-only"
classifiers = [
    "Topic :: Security",
]
packages = [
    { include = "witnessme"}
]

[tool.poetry.scripts]
witnessme = 'witnessme.console.witnessme:run'
wmapi = 'witnessme.console.wmapi:run'
wmdb = 'witnessme.console.wmdb:run'

[tool.poetry.dependencies]
python = "^3.7.0"
fastapi = "^0.55.1"
xmltodict = "^0.12.0"
terminaltables = "^3.1.0"
imgcat = "^0.5.0"
pyyaml = "^5.3.1"
aiosqlite = "^0.13.0"
prompt-toolkit = "^3.0.5"
pydantic = "^1.5.1"
uvicorn = "^0.11.5"
pyppeteer = "^0.2.2"
python-multipart = "^0.0.5"
jinja2 = "^2.11.2"
aiodns = "^2.0.0"
lxml = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
flake8 = "^3.8.2"
pylint = "^2.5.2"
black = "^19.10b0"
requests = "^2.23.0"
pytest-asyncio = "^0.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
